mv CustomOnChangeProps and AutocompleteNewValue to @/common
