+++ packages/rhf-mui-components/src/common/types/components.ts
@@ -15,3 +15,23 @@ export type OptionValue<
         ? Option[ValueKey]
         : never
       : never;
+
+/**
+ * RHF field value for Autocomplete components mirrors
+ * MUI `AutocompleteValue<string, Multiple, DisableClearable, false>` for primitives.
+ * Tuple checks avoid distributive `boolean` breaking the conditional.
+ */
+export type AutocompleteNewValue<
+  Multiple extends boolean,
+  DisableClearable extends boolean
+> = [Multiple] extends [true]
+  ? [DisableClearable] extends [true]
+    ? string[]
+    : string[] | null
+  : [DisableClearable] extends [true]
+    ? string
+    : string | null;
+
+export type CustomOnChangeProps<T, V> = T & {
+  rhfOnChange: (value: V) => void;
+};
