+++ packages/rhf-mui-components/src/common/types/index.ts
@@ -1,2 +1,2 @@
-export * from './components';
+export type * from './components';
 export type * from './mui';
