+++ packages/rhf-mui-components/src/mui/autocomplete-object/index.tsx
@@ -17,7 +17,8 @@ import {
   type Control,
   type RegisterOptions
 } from 'react-hook-form';
-import Autocomplete, {
+import Autocomplete,
+{
   type AutocompleteProps,
   type AutocompleteChangeDetails,
   type AutocompleteChangeReason,
@@ -35,9 +36,10 @@ import {
   type FormLabelProps,
   type FormHelperTextProps,
   type AutoCompleteTextFieldProps,
-  type MuiChipProps
+  type MuiChipProps,
+  type CustomOnChangeProps
 } from '@/common';
-import type { KeyValueOption, CustomOnChangeProps, CustomComponentIds } from '@/types';
+import type { KeyValueOption, CustomComponentIds } from '@/types';
 import {
   fieldNameToLabel,
   useFieldIds,
