+++ packages/rhf-mui-components/src/types/common.ts
@@ -12,23 +12,3 @@ export type CustomComponentIds = Partial<{
   helperText: string;
   error: string;
 }>;
-
-export type CustomOnChangeProps<T, V> = T & {
-  rhfOnChange: (value: V) => void;
-};
-
-/**
- * RHF field value for Autocomplete components mirrors
- * MUI `AutocompleteValue<string, Multiple, DisableClearable, false>` for primitives.
- * Tuple checks avoid distributive `boolean` breaking the conditional.
- */
-export type AutocompleteNewValue<
-  Multiple extends boolean,
-  DisableClearable extends boolean
-> = [Multiple] extends [true]
-  ? [DisableClearable] extends [true]
-    ? string[]
-    : string[] | null
-  : [DisableClearable] extends [true]
-    ? string
-    : string | null;
